backend/requirements.txt
fastapi
pydantic
pandas
langchain
langchain-groq
chromadb
sentence-transformers
python-dotenv